Project-OLEG-Telegram-AI-Personality-Bot/pyproject.toml
[tool.poetry]
name = "oleg-bot"
version = "8.0.0"
description = "Цифровой гигачад — Telegram-бот с ИИ, модерацией и игровыми механиками"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "app" },
]
repository = "https://github.com/your-repo/oleg-bot"
keywords = ["telegram", "bot", "aiogram", "ollama", "ai"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: Russian",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "*"
python-dotenv = "*"
SQLAlchemy = "*"
aiosqlite = "*"
alembic = "*"
httpx = "*"
APScheduler = "*"
pytz = "*"
python-dateutil = "*"
uvloop = { version = "*", markers = "sys_platform != 'win32'" }
orjson = "*"
Pillow = "*"
yt-dlp = "*"
chromadb = "*"
ollama = "*"
pydantic-settings = "*"
arq = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
black = "*"
flake8 = "*"
isort = "*"
mypy = "*"

[tool.poetry.scripts]
oleg = "app.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
